Re #7 remove rebalance_cb; use assign to set start offsets insead
